+++ src/pb/parser.h
@@ -2,7 +2,6 @@
 #define PB_PARSER_H
 
 #include<stdio.h>
-#include "data/int64.h"
 
 typedef unsigned long long uint64;
 
@@ -11,13 +10,13 @@ typedef unsigned long long uint64;
  * @param stream The stream to read from
  * @param limit Maximum number of bytes to read
  */
-void readObjectFromStream(FILE* stream, unsigned long long limit);
+void readObjectFromStream(FILE* stream, uint64 limit);
 
 /**
  * Reads a 64-bit number from the stream
  * @param stream The stream to read from
  */
-int64 readFixed64(FILE* stream);
+uint64 readFixed64(FILE* stream);
 
 /**
  * Reads a varint from the stream and returns as int64
